+++ thunderstore_package/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Whats changed
 
+# v.1.1.1
+- As you have long asked me, I have added support for changing brightness and range via the LetalConfig mod. I will be glad to get a star on GitHub and feedback on bugs!
+- Now you can change range or/and intensity of helmet and flashlights through LethalConfig without restart
+- Your config not flushed after game close/reboot
+- Maximum intensity and range of light can be set to 2x extends default values
+- Default values dont changed in this update
+
 # v.1.1.0
 - Set upper values for power of Helmet Light - because this use another values on start